+++ App_Estudios_G7/App_Estudios_G7.Tests/App.config
@@ -9,7 +9,7 @@
     </appSettings>
   
     <connectionStrings>
-      <add name="Sistema_estudiosEntities" connectionString="metadata=res://*/Models.Modelo.csdl|res://*/Models.Modelo.ssdl|res://*/Models.Modelo.msl;provider=System.Data.SqlClient;provider connection string=&quot;data source=SERGIO\SERGIO;initial catalog=Sistema_estudios;integrated security=True;MultipleActiveResultSets=True;App=EntityFramework&quot;" providerName="System.Data.EntityClient" />
+      <add name="Sistema_estudiosEntities" connectionString="metadata=res://*/Models.Modelo.csdl|res://*/Models.Modelo.ssdl|res://*/Models.Modelo.msl;provider=System.Data.SqlClient;provider connection string=&quot;data source=LAPTOP-IFGR27P8;initial catalog=Sistema_estudios;integrated security=True;MultipleActiveResultSets=True;App=EntityFramework&quot;" providerName="System.Data.EntityClient" />
     </connectionStrings>
   <runtime>
     <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
