+++ App_Estudios_G7/App_Estudios_G7.Tests/Controllers/HomeControllerTest.cs
@@ -25,18 +25,7 @@ namespace App_Estudios_G7.Tests.Controllers
             Assert.IsNotNull(result);
         }
 
-        [TestMethod]
-        public void About()
-        {
-            // Arrange
-            HomeController controller = new HomeController();
-
-            // Act
-            ViewResult result = controller.About() as ViewResult;
-
-            // Assert
-            Assert.AreEqual("Your application description page.", result.ViewBag.Message);
-        }
+        
 
         [TestMethod]
         public void Contact()
@@ -49,6 +38,48 @@ namespace App_Estudios_G7.Tests.Controllers
 
             // Assert
             Assert.IsNotNull(result);
+        }
+
+        [TestMethod]
+        public void LoginUsuarioNoNull()
+        {
+            // Arrange
+            CuentaController controller = new CuentaController();
+
+            // Act
+            ViewResult result = controller.Login() as ViewResult;
+
+            // Assert
+            Assert.IsNotNull(result);
+
+        }
+
+        [TestMethod]
+        public void RegistrarUsuarioNoNull()
+        {
+            // Arrange
+            CuentaController controller = new CuentaController();
+
+            // Act
+            ViewResult result = controller.Registrarse() as ViewResult;
+
+            // Assert
+            Assert.IsNotNull(result);
+
+        }
+
+        [TestMethod]
+        public void VerificarRegistroNoNull()
+        {
+            // Arrange
+            CuentaController controller = new CuentaController();
+
+            // Act
+            bool result = controller.VerificarRegistro("usuario");
+
+            // Assert
+            Assert.IsNotNull(result);
+
         }
     }
 }
