+++ App_Estudios_G7/App_Estudios_G7.Tests/Controllers/RegistrarseControllerTest.cs
@@ -2,11 +2,12 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
-using App_Estudios_G7.Models;
-using App_Estudios_G7.Controllers;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Web.Mvc;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using App_Estudios_G7;
+using App_Estudios_G7.Controllers;
+using System.Data;
+using App_Estudios_G7.Models;
 
 namespace App_Estudios_G7.Tests.Controllers
 {
@@ -27,33 +28,7 @@ namespace App_Estudios_G7.Tests.Controllers
         
         }
 
-        [TestMethod]
-        public void RegistrarUsuarioNoNull()
-        {
-            // Arrange
-            CuentaController controller = new CuentaController();
-
-            // Act
-            ViewResult result = controller.Registrarse() as ViewResult;
-
-            // Assert
-            Assert.IsNotNull(result);
-
-        }
-
-        [TestMethod]
-        public void VerificarRegistroNoNull()
-        {
-            // Arrange
-            CuentaController controller = new CuentaController();
-
-            // Act
-            bool result = controller.VerificarRegistro("usuario");
-
-            // Assert
-            Assert.IsNotNull(result);
-
-        }
+        
 
     }
 }
