+++ App_Estudios_G7/App_Estudios_G7/Views/Curso/ContenidoCurso.cshtml
@@ -54,7 +54,7 @@
 
     @*Boton de creacion de tarea*@
     <div class="row" style="text-align:left;">
-        <a class="btn btn-primary btn-lg" href="/Curso/CrearCurso">Crear tarea</a>
+        <a class="btn btn-primary btn-lg" href="/Tarea/TareaCreate">Crear tarea</a>
     </div>
 
     @*Tablas de contenido del curso*@
