+++ App_Estudios_G7/App_Estudios_G7/Views/Curso/MisCursos.cshtml
@@ -1,5 +1,5 @@
-﻿@using App_Estudios_G7.Models
-@model List<App_Estudios_G7.Models.ListaCursos>
+﻿@using App_Estudios_G7.Models.ViewModels
+@model List<App_Estudios_G7.Models.ViewModels.ListaCursos>
 
 @{
     ViewBag.Title = "Mis Cursos";
