+++ App_Estudios_G7/App_Estudios_G7/Views/Tarea/TareaCreate.cshtml
@@ -33,10 +33,10 @@
     </div>
 
     <div class="form-group">
-        @Html.LabelFor(model => model.Ponderacion, htmlAttributes: new { @class = "control-label col-md-2" })
+        @Html.LabelFor(model => model.ponderacion, htmlAttributes: new { @class = "control-label col-md-2" })
         <div class="col-md-10">
-            @Html.EditorFor(model => model.Ponderacion, new { htmlAttributes = new { @class = "form-control" } })
-            @Html.ValidationMessageFor(model => model.Ponderacion, "", new { @class = "text-danger" })
+            @Html.EditorFor(model => model.ponderacion, new { htmlAttributes = new { @class = "form-control" } })
+            @Html.ValidationMessageFor(model => model.ponderacion, "", new { @class = "text-danger" })
         </div>
     </div>
